- Refactor item lists to new class tree:         - FileListTableViewController                 - SharesFilelistTableViewController         - QueryFileListTableViewController                 - ClientQueryViewController                 - CustomFileListTableViewController - Restructure source file locations
