+++ ownCloud/Client/Viewer/DisplayHostViewController.swift
@@ -37,6 +37,7 @@ class DisplayHostViewController: UIPageViewController {
 		}
 	}
 	private var query: OCQuery
+	private var queryStarted : Bool = false
 	private weak var viewControllerToTansition: DisplayViewController?
 	private var queryObservation : NSKeyValueObservation?
 
@@ -50,6 +51,11 @@ class DisplayHostViewController: UIPageViewController {
 
 		super.init(transitionStyle: .scroll, navigationOrientation: .horizontal, options: nil)
 
+		if query.state == .stopped {
+			core.start(query)
+			queryStarted = true
+		}
+
 		queryObservation = query.observe(\OCQuery.hasChangesAvailable, options: [.initial, .new]) { [weak self] (query, _) in
 			query.requestChangeSet(withFlags: .onlyResults) { ( _, changeSet) in
 				guard let changeSet = changeSet  else { return }
@@ -58,6 +64,7 @@ class DisplayHostViewController: UIPageViewController {
 				}
 			}
 		}
+
 		Theme.shared.register(client: self)
 	}
 
@@ -66,6 +73,11 @@ class DisplayHostViewController: UIPageViewController {
 	}
 
 	deinit {
+		if queryStarted {
+			core?.stop(query)
+			queryStarted = false
+		}
+
 		queryObservation?.invalidate()
 		Theme.shared.unregister(client: self)
 	}
