+++ ownCloud/FileLists/FileListTableViewController.swift
@@ -1,0 +1,202 @@
+//
+//  FileListTableViewController.swift
+//  ownCloud
+//
+//  Created by Matthias Hühne on 21.05.19.
+//  Copyright © 2019 ownCloud GmbH. All rights reserved.
+//
+
+import UIKit
+import ownCloudSDK
+
+class FileListTableViewController: UITableViewController, ClientItemCellDelegate, Themeable {
+	weak var core : OCCore?
+
+	let estimatedTableRowHeight : CGFloat = 80
+
+	var progressSummarizer : ProgressSummarizer?
+	private var _actionProgressHandler : ActionProgressHandler?
+
+	public init(core inCore: OCCore) {
+		core = inCore
+		super.init(style: .plain)
+
+		progressSummarizer = ProgressSummarizer.shared(forCore: inCore)
+	}
+
+	required init?(coder aDecoder: NSCoder) {
+		fatalError("init(coder:) has not been implemented")
+	}
+
+	deinit {
+		Theme.shared.unregister(client: self)
+	}
+
+	func makeActionProgressHandler() -> ActionProgressHandler {
+		if _actionProgressHandler == nil {
+			_actionProgressHandler = { [weak self] (progress, publish) in
+				if publish {
+					self?.progressSummarizer?.startTracking(progress: progress)
+				} else {
+					self?.progressSummarizer?.stopTracking(progress: progress)
+				}
+			}
+		}
+
+		return _actionProgressHandler!
+	}
+
+	// MARK: - Item retrieval
+	func item(for cell: ClientItemCell) -> OCItem? {
+		return cell.item
+	}
+
+	func itemAt(indexPath : IndexPath) -> OCItem? {
+		return (self.tableView.cellForRow(at: indexPath) as? ClientItemCell)?.item
+	}
+
+	// MARK: - ClientItemCellDelegate
+	func moreButtonTapped(cell: ClientItemCell) {
+		guard let item = self.item(for: cell), let core = core else {
+			return
+		}
+
+		let actionsLocation = OCExtensionLocation(ofType: .action, identifier: .moreItem)
+		let actionContext = ActionContext(viewController: self, core: core, items: [item], location: actionsLocation)
+
+		if let moreViewController = Action.cardViewController(for: item, with: actionContext, progressHandler: makeActionProgressHandler()) {
+			self.present(asCard: moreViewController, animated: true)
+		}
+	}
+
+	// MARK: - Themable
+	func applyThemeCollection(theme: Theme, collection: ThemeCollection, event: ThemeEvent) {
+		self.tableView.applyThemeCollection(collection)
+
+		if event == .update {
+			self.reloadTableData()
+		}
+	}
+
+	// MARK: - Visibility handling
+	private var viewControllerVisible : Bool = false
+
+	override func viewWillDisappear(_ animated: Bool) {
+		super.viewWillDisappear(animated)
+
+		viewControllerVisible = false
+	}
+
+	override func viewWillAppear(_ animated: Bool) {
+		super.viewWillAppear(animated)
+
+		viewControllerVisible = true
+		self.reloadTableData(ifNeeded: true)
+	}
+
+	// MARK: - View setup
+	override func viewDidLoad() {
+		super.viewDidLoad()
+
+		self.navigationController?.navigationBar.prefersLargeTitles = false
+		Theme.shared.register(client: self, applyImmediately: true)
+		self.tableView.estimatedRowHeight = estimatedTableRowHeight
+
+		self.registerCellClasses()
+	}
+
+	func registerCellClasses() {
+		self.tableView.register(ClientItemCell.self, forCellReuseIdentifier: "itemCell")
+	}
+
+	// MARK: - Reload Data
+	private var tableReloadNeeded = false
+
+	func reloadTableData(ifNeeded: Bool = false) {
+		/*
+			This is a workaround to cope with the fact that:
+			- UITableView.reloadData() does nothing if the view controller is not currently visible (via viewWillDisappear/viewWillAppear), so cells may hold references to outdated OCItems
+			- OCQuery may signal updates at any time, including when the view controller is not currently visible
+
+			This workaround effectively makes sure reloadData() is called in viewWillAppear if a reload has been signalled to the tableView while it wasn't visible.
+		*/
+		if !viewControllerVisible {
+			tableReloadNeeded = true
+		}
+
+		if !ifNeeded || (ifNeeded && tableReloadNeeded) {
+			self.tableView.reloadData()
+
+			if viewControllerVisible {
+				tableReloadNeeded = false
+			}
+
+			self.restoreSelectionAfterTableReload()
+		}
+	}
+
+	func restoreSelectionAfterTableReload() {
+	}
+
+	// MARK: - Table view delegate
+	var lastTappedItemLocalID : String?
+
+	func query(forItem: OCItem) -> OCQuery? {
+		if let path = forItem.path {
+			return OCQuery(forPath: path)
+		}
+
+		return nil
+	}
+
+	override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+		if !self.tableView.isEditing {
+			guard let rowItem : OCItem = itemAt(indexPath: indexPath) else {
+				return
+			}
+
+			if let core = self.core {
+				switch rowItem.type {
+					case .collection:
+						if let path = rowItem.path {
+							self.navigationController?.pushViewController(ClientQueryViewController(core: core, query: OCQuery(forPath: path)), animated: true)
+						}
+
+					case .file:
+						guard let query = self.query(forItem: rowItem) else {
+							return
+						}
+
+						if lastTappedItemLocalID != rowItem.localID {
+							lastTappedItemLocalID = rowItem.localID
+
+							if let progress = core.downloadItem(rowItem, options: [ .returnImmediatelyIfOfflineOrUnavailable : true ], resultHandler: { [weak self, query] (error, core, item, _) in
+
+								guard let self = self else { return }
+								OnMainThread { [weak core] in
+									if (error == nil) || (error as NSError?)?.isOCError(withCode: .itemNotAvailableOffline) == true {
+										if let item = item, let core = core {
+											if item.localID == self.lastTappedItemLocalID {
+												let itemViewController = DisplayHostViewController(core: core, selectedItem: item, query: query)
+												itemViewController.hidesBottomBarWhenPushed = true
+												itemViewController.progressSummarizer = self.progressSummarizer
+												self.navigationController?.pushViewController(itemViewController, animated: true)
+											}
+										}
+									}
+
+									if self.lastTappedItemLocalID == item?.localID {
+										self.lastTappedItemLocalID = nil
+									}
+								}
+							}) {
+								progressSummarizer?.startTracking(progress: progress)
+							}
+						}
+				}
+			}
+
+			tableView.deselectRow(at: indexPath, animated: true)
+		}
+	}
+}
